Move WidgetChromeActionableArea to widgets
