+++ routes/dashboard/widget-dashboard/components/widget-chrome/index.ts
@@ -1,1 +1,1 @@
-export { WidgetChrome, WidgetChromeActionableArea } from './widget-chrome';
+export { WidgetChrome } from './widget-chrome';
