+++ routes/dashboard/widget-dashboard/components/widgets/widget-chrome-actionable-area.module.css
@@ -1,0 +1,6 @@
+.widgetChromeActionableArea {
+	position: absolute;
+	top: var(--wpds-dimension-padding-md);
+	inset-inline-end: var(--wpds-dimension-padding-md);
+	z-index: 1;
+}
